Fix qtlocation build for WEC7.
Apparently math.h in WEC7 does not define M_PI used by qtlocation.
Added the requiresd define the necessary files.

qSqrt returns qreal and it is defined to 'float' on some configurations
and to 'double' in other configurations. In those configirations where
it is defined to float, no suitable overload for division is found.
Added explicit cast for qSqrt return value to double to fix the build.

+++ src/location/maps/qdoublevector2d.cpp
@@ -70,7 +70,7 @@ QDoubleVector2D QDoubleVector2D::normalized() const
     if (qFuzzyIsNull(len - 1.0))
         return *this;
     else if (!qFuzzyIsNull(len))
-        return *this / qSqrt(len);
+        return *this / (double)qSqrt(len);
     else
         return QDoubleVector2D();
 }
