+++ src/location/maps/qdoublevector3d.cpp
@@ -69,7 +69,7 @@ QDoubleVector3D QDoubleVector3D::normalized() const
     if (qFuzzyIsNull(len - 1.0))
         return *this;
     else if (!qFuzzyIsNull(len))
-        return *this / qSqrt(len);
+        return *this / (double)qSqrt(len);
     else
         return QDoubleVector3D();
 }
